+++ src/components/Controllers/Table/store.ts
@@ -7,7 +7,7 @@ import { createStore } from 'zustand/vanilla'
 import { createSelector } from 'reselect'
 import { assert } from 'ts-essentials'
 import { Client } from '../../../client'
-import { ITableEditor } from '../../Editors/Table'
+import { ITableEditor } from '../../Parts/Table'
 import { TableProps } from './index'
 import * as settings from '../../../settings'
 import * as helpers from '../../../helpers'
