+++ src/components/Parts/Table/settings.ts
